requirements.txt
anyio==3.6.1
autopep8==1.6.0
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cssselect2==0.6.0
fastapi==0.78.0
fonttools==4.34.4
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
httplib2==0.20.4
httptools==0.4.0
idna==3.3
lxml==4.9.1
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.2.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pydyf==0.2.0
pyparsing==3.0.9
pyphen==0.12.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tinycss2==1.1.1
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
webencodings==0.5.1
websockets==10.3
zopfli==0.2.1
